% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{closest_index}
\alias{closest_index}
\title{Computes the index that places the value in the vector}
\usage{
closest_index(vec, val)
}
\arguments{
\item{vec}{vector of numeric values to search}

\item{val}{the val to match into vec}

\item{1-based}{index of the closest match in vec}
}
\description{
This is not like \code{findInterval} which places by order.  Here order 
is irrelevant.
}

