% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib1.R
\docType{class}
\name{Grib1RefClass-class}
\alias{Grib1RefClass-class}
\alias{Grib1RefClass}
\title{Grib1RefClass}
\description{
Grib1RefClass
}
\section{Fields}{

\describe{
\item{\code{filename}}{the name of the GRIB1 file}

\item{\code{inv}}{a tibble of the GRIB1 inventory}
}}

\section{Methods}{

\describe{
\item{\code{read_band(..., bb = NULL)}}{Read a raster band}

\item{\code{read_grib1(band = select_band(.self$inv), bb = NULL)}}{Read a raster band}
}}

