% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAM218.R
\name{has_var}
\alias{has_var}
\title{Test if variable names include the ones specified}
\usage{
has_var(nc, name, ignore.case = TRUE)
}
\arguments{
\item{nc}{a ncdf4 class object}

\item{name}{character, one or more names to match}

\item{ignore.case}{logical it TRUE ignore case in the pattern}
}
\value{
logical vector where TRUE indicates a match is found
}
\description{
Test if variable names include the ones specified
}
