% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam218-package.R
\docType{package}
\name{nam218-package}
\alias{nam218-package}
\title{A simplified interface to NOMADS NAM 218 forecasts and archives.}
\description{
A simplified interface to NOMADS NAM 218 forecasts and archives.
}
\examples{
 \dontrun{
 url = "http://nomads.ncdc.noaa.gov/thredds/dodsC/namanl/200606/20060601/namanl_218_20060601_0000_000.grb"
 X <- NAM218(url)
 RH <- X$get_layer("Relative_humidity")
 SWG <- X$get_layer("Surface_wind_gust")
}
}

