% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAM218.R
\name{NAM218}
\alias{NAM218}
\title{Create an instance of NAM218RefClass}
\usage{
NAM218(nc)
}
\arguments{
\item{nc}{either a url, ncdf4 object, DatasetsRefClass object or null}
}
\value{
NAM218RefClass object or NULL
}
\description{
Create an instance of NAM218RefClass
}

