% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{nam218_download}
\alias{nam218_download}
\title{Download a GRIB file}
\usage{
nam218_download(X, dest = NULL)
}
\arguments{
\item{X}{DatasetsRefClass object}

\item{dest}{the destination file name, if not provided then it is borrowed from X}
}
\value{
value returned by utils::download.file()
}
\description{
Download a GRIB file
}
