% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nam218_url_error}
\alias{nam218_url_error}
\title{Test if a NAM218 OpenDAP URL points to accessible datasets.}
\usage{
nam218_url_error(x)
}
\arguments{
\item{x}{character URL or DatasetsRefClass object}
}
\value{
logical TRUE if the URL fails to point to an available dataset
}
\description{
Test if a NAM218 OpenDAP URL points to accessible datasets.
}
