% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAM218.R
\name{NAM218RefClass_bb_to_xy}
\alias{NAM218RefClass_bb_to_xy}
\title{Convert from longlat bounding box to start, count for x and y}
\arguments{
\item{bb}{SpatialPolygons object to project to Lamber Conformal Conic}
}
\value{
list of start[x,y] and count[x,y], bbox (original), ext (Extent coordinates)
}
\description{
Convert from longlat bounding box to start, count for x and y
}

