% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAM218.R
\name{NAM218RefClass_get_inventory}
\alias{NAM218RefClass_get_inventory}
\title{Retrieve the inventory for the grib file}
\arguments{
\item{form}{character either 'text' for a character vector of 'data.frame'}
}
\value{
NULL, character vector or data.frame
}
\description{
Retrieve the inventory for the grib file
}

