% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAM218.R
\name{NAM218RefClass_get_layer}
\alias{NAM218RefClass_get_layer}
\title{Retrieve a parameter}
\arguments{
\item{name}{the name of the layer}

\item{bb}{a 4-element vector of longlat as [xmin,xmax,ymin,ymax] defining the region of interest
If NULL then extent of the BB element of the  NAM218RefClass object is used}
}
\value{
RasterLayer object or NULL
}
\description{
Retrieve a parameter
}

