% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAM218.R
\name{NAM218RefClass_RasterLayer}
\alias{NAM218RefClass_RasterLayer}
\title{Convert a matrix into a RasterLayer object}
\arguments{
\item{x}{matrix}

\item{name}{character name of the layer}

\item{flip_it}{character ('y' the default, 'x' or '') for flip-y, flip-x or no-flip}
}
\value{
RasterLayer object
}
\description{
Convert a matrix into a RasterLayer object
}

