% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{nam_nc_open}
\alias{nam_nc_open}
\title{Retrieve a ncdf4 object - only if ncdf4 package is available}
\usage{
nam_nc_open(X)
}
\arguments{
\item{X}{DatasetsRefClass object}
}
\value{
value returned by ncdf4::nc_open()
}
\description{
Retrieve a ncdf4 object - only if ncdf4 package is available
}

