% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{nam_query}
\alias{nam_query}
\title{Run a query for NAM-218 datasets}
\usage{
nam_query(what = c("analysis", "forecast")[1], day = c("20060601",
  format(as.POSIXct(Sys.time(), format = "\%Y\%m\%d")))[1], time = c(NA,
  "0000", "0600", "1200", "1800")[3], ahead = c(NA, "000", "003", "006",
  "084")[1])
}
\arguments{
\item{what}{character specifies 'analysis' or 'forecast'}

\item{day}{either POSIXct or an 8 character date (YYYYmmdd)}

\item{time}{4 character forecast period time stamp ('0000', '0006', etc.)}

\item{ahead}{3 character cycle timestamp ('000' now cast, '003 three hours ahead, etc)}
}
\value{
DatasetsRefClass or NULL
}
\description{
Run a query for NAM-218 datasets
}
