% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMANL.R
\name{NAMANL}
\alias{NAMANL}
\title{Create an instance of NAMANLRefClass}
\usage{
NAMANL(nc, proj = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{nc}{either a path, url, ncdf4 object or null}

\item{proj}{projection string for automatic transformation from NAM-ANL's native 'lcc' projection}
}
\value{
NAMANLRefClass object
}
\description{
Create an instance of NAMANLRefClass
}

