% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{namanl_browse}
\alias{namanl_browse}
\title{Browse the OPeNDAP resources online}
\usage{
namanl_browse(X, what = c("opendap_form", "netcdfserver")[1])
}
\arguments{
\item{X}{DatasetsRefClass object}

\item{what}{character, which format do you prefer?}
}
\value{
the value of httr::BROWSE
}
\description{
Browse the OPeNDAP resources online
}

