% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib1.R
\name{namanl_grib1_https}
\alias{namanl_grib1_https}
\title{Fetch a grb and companion inv file via https}
\usage{
namanl_grib1_https(
  name = "namanl_218_20040303_0000_000",
  path = tempdir(),
  what = c("grb", "inv")[1],
  ...
)
}
\arguments{
\item{name}{the name of the file to retrieve in namanl_218_YYYYmmdd_ffff_aaa pattern}

\item{path}{the destination path to save the file}

\item{what}{character vector of  grb, inv or both}

\item{...}{further arguments for \code{namanl_grib1_ftp_uri()}}
}
\value{
character vector of the filenames downloaded.
 \itemize{
     \item{zero elements means nothing downloaded}
     \item{one element means one file downloaded}
     \item{two elements mean both grb and inv files downloaded}
 }
}
\description{
Fetch a grb and companion inv file via https
}
