% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib1.R
\name{namanl_grib1_https_uri}
\alias{namanl_grib1_https_uri}
\title{Construct a https uri given the name of a GRIB1 file}
\usage{
namanl_grib1_https_uri(name = "namanl_218_20040303_0000_000",
  root = "https://nomads.ncdc.noaa.gov/data/namanl")
}
\arguments{
\item{name}{character, the name stub in the form of namanl_218_YYYYmmdd_ffff_aaa
where ffff is the forecast statement time and aaa is the forecast hours ahead of the statement}

\item{root}{the root uri}
}
\value{
two element vector of *.grb and *.inv uri
}
\description{
Construct a https uri given the name of a GRIB1 file
}
