% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib1.R
\name{namanl_grib1_name_parse}
\alias{namanl_grib1_name_parse}
\title{Parse a NAMANL grib filename}
\usage{
namanl_grib1_name_parse(name = "namanl_218_20040303_0000_000")
}
\arguments{
\item{name}{the name in 'namanl_218_YYYYmmdd_ffff_aaa' pattern}
}
\value{
named character vector
\itemize{
  \item{name the original name stripped of path and extension}
  \item{product namanl_218 (fist two parts joined with "_")}
  \item{ymd YYYYmmdd}
  \item{ym  YYYYmm}
  \item{y YYYY}
  \item{m mm}
  \item{d dd}
  \item{ftime  forecast statement time as 24h}
  \item{atime  forecast hours ahead relative to statement time}
}
}
\description{
Parse a NAMANL grib filename
}
