% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{namanl_query}
\alias{namanl_query}
\title{Run a query for NAM-ANL datasets}
\usage{

  namanl_query(top = "http://nomads.ncdc.noaa.gov/thredds/catalog/namanl/catalog.xml",
  day = c("20060601", format(as.POSIXct(Sys.time(), format =
  "\%Y\%m\%d")))[1], time = c("0000", "0600", "1200", "1800")[3],
  ahead = c("000", "003", "006")[1], stub = "namanl_218")
}
\arguments{
\item{top}{character, the top level uri - XML}

\item{day}{either POSIXct or an 8 character date (YYYYmmdd)}

\item{time}{4 character forecast period time stamp ('0000', '0006', etc.)}

\item{ahead}{3 character cycle timestamp ('000' now cast, '003 three hours ahead, etc)}
}
\value{
DatasetsRefClass or NULL
}
\description{
Run a query for NAM-ANL datasets
}

