% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMANL.R
\name{namanl_Relative_humidity}
\alias{namanl_Relative_humidity}
\title{Retrieve variable Relative_humidity [x,y,isobaric,time] as a matrix}
\usage{
namanl_Relative_humidity(X, isobar = 1000)
}
\arguments{
\item{X}{NAMANLRefClass object}

\item{isobar}{the isobaric lavel to match (we take the closest match)}
}
\value{
numeric matrix
}
\description{
Retrieve variable Relative_humidity [x,y,isobaric,time] as a matrix
}

