% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMANL.R
\name{namanl_Surface_wind_gust}
\alias{namanl_Surface_wind_gust}
\title{Retrieve variable Surface_wind_gust[x,y,time] as a matrix}
\usage{
namanl_Surface_wind_gust(X, ...)
}
\arguments{
\item{X}{NAMANLRefClass object}

\item{...}{further arguments (unused)}
}
\value{
numeric matrix
}
\description{
Retrieve variable Surface_wind_gust[x,y,time] as a matrix
}

