% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam_anl.R
\name{namanl_threshold_date}
\alias{namanl_threshold_date}
\title{Retrieve the NAM-ANL inclusive threshold that defines 'recent' data}
\usage{
namanl_threshold_date()
}
\value{
Date class object
}
\description{
Retrieve the NAM-ANL inclusive threshold that defines 'recent' data
}
