% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMANL.R
\name{namanl_Total_precipitation}
\alias{namanl_Total_precipitation}
\title{Retrieve variable Total_precipitation [x,y,time1] as a matrix}
\usage{
namanl_Total_precipitation(X, ...)
}
\arguments{
\item{X}{NAMANLRefClass object}

\item{...}{further arguments (unused)}
}
\value{
numeric matrix
}
\description{
Retrieve variable Total_precipitation [x,y,time1] as a matrix
}

