% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namanl.R
\name{namanl_url}
\alias{namanl_url}
\title{Construct thredds base urls for provided date(s)}
\usage{
namanl_url(dates = "2018-12-18", base_url = namanl_base_url(),
  ftime = c("0000", "0600", "1200", "1800")[1], ahead = c("000", "003",
  "006", "084")[1])
}
\arguments{
\item{dates}{Date-class, one or more dates}

\item{base_url}{character, the base URL}

\item{ftime}{4 character forecast period time stamp ('0000', '0006', etc.)}

\item{ahead}{3 character cycle timestamp ('000' now cast, '003 three hours ahead, etc)}
}
\value{
character URLs
}
\description{
Completed URLs have the form
"https://www.ncdc.noaa.gov/thredds/dodsC/namanl/201812/20181218/namanl_218_20181218_0000_000.grb2"
which is the same as                  "base_url/YYYYmm/YYYYmmdd/namanl_218_YYYYmmdd_ffff_aaa.grb2"
}
