% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMANL.R
\name{NAMANLRefClass_get_layer}
\alias{NAMANLRefClass_get_layer}
\title{Retrieve a parameter}
\arguments{
\item{name}{the name of the layer}
}
\value{
RasterLayer object or NULL
}
\description{
Retrieve a parameter
}

