% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMcast.R
\name{NAMcast}
\alias{NAMcast}
\title{Create an instance of NAMcastRefClass}
\usage{
NAMcast(
  nc = format(Sys.Date(),
    "http://nomads.ncep.noaa.gov:80/dods/nam/nam\%Y\%m\%d/nam_00z")
)
}
\arguments{
\item{nc}{either a url, ncdf4 object or null}
}
\value{
NAMcastRefClass object or NULL
}
\description{
Create an instance of NAMcastRefClass
}
