% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam_cast.R
\name{namcast_open}
\alias{namcast_open}
\title{Open an ncdf4 connection to an NAM(cast) uri}
\usage{
namcast_open(uri, verbose = FALSE)
}
\arguments{
\item{uri}{character the URL}

\item{verbose}{logical, if TRUE voice warnings}
}
\value{
either ncdf4 object or NULL
}
\description{
Open an ncdf4 connection to an NAM(cast) uri
}
