% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam_cast.R
\name{namcast_parse_filename}
\alias{namcast_parse_filename}
\title{Parse NAM forecast filenames}
\usage{
namcast_parse_filename(
  x = c("nam_218_20210105_0000_084.grb2", "nam_218_20210105_0000_081.grb2",
    "nam_218_20210105_0000_078.grb2")
)
}
\arguments{
\item{x}{character, one or more filenames}
}
\value{
tibble of filename constituent parts
\itemize{
 \item{model most likely "nam"}
 \item{domain character, code for the extent/grid see \href{https://www.nco.ncep.noaa.gov/pmb/docs/on388/tableb.html}{NCEP STORAGE GRIDS} 
 \item{date Date class}
 \item{ftime character, 4 digit integer code of forecast statement hour (aka cycle time)}
 \item{ahead character, 3 digit integer code of the forecasted hour}}
}
}
\description{
Filenames such as "nam_218_20210105_0000_069.grb2"
are comprised of "product_YYYYmmdd_ffff_aaa.ext"
}
