% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam_cast.R
\name{namcast_threshold_date}
\alias{namcast_threshold_date}
\title{Retrieve the NAM-CAST inclusive threshold that defines 'recent' data}
\usage{
namcast_threshold_date()
}
\value{
Date class object
}
\description{
Retrieve the NAM-CAST inclusive threshold that defines 'recent' data
}
