% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMcast.R
\name{NAMcastRefClass_get_time}
\alias{NAMcastRefClass_get_time}
\title{Compute the POSIXct timestamps}
\arguments{
\item{shift}{numeric, the raw day numbers and epoch conspire to cause a two day 
difference from what is expected. So this param is used to correct that.}
}
\value{
POSIXct vector or NULL
}
\description{
Compute the POSIXct timestamps
}
\seealso{
\url{https://github.com/dankelley/oce/issues/738}

\url{http://www.epic.noaa.gov/java/ncBrowse/mail/msg00059.html}
}

