% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{ncdim_list}
\alias{ncdim_list}
\title{Retrieve a vector of pretty dimensions}
\usage{
ncdim_list(NC)
}
\arguments{
\item{NC}{a ncdf4 object}
}
\value{
a named vector of dimesnions
}
\description{
Retrieve a vector of pretty dimensions
}

