% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{ncvar_dim_name}
\alias{ncvar_dim_name}
\title{Retrieve a named vector of variable dimensions as a string.  If a variable
 has dimensions [x, y, isobaric, time] then the value will be
 'x_y_isobaric_time'.}
\usage{
ncvar_dim_name(nc, sep = "_")
}
\arguments{
\item{sep}{character used to separate the dimension names}

\item{NC}{a ncdf4 object}
}
\value{
a named vector of variable dimension names
}
\description{
Retrieve a named vector of variable dimensions as a string.  If a variable
 has dimensions [x, y, isobaric, time] then the value will be
 'x_y_isobaric_time'.
}
