% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{ncvar_info}
\alias{ncvar_info}
\title{Retrieve a matrix generic information about each variable}
\usage{
ncvar_info(NC, sep = " ")
}
\arguments{
\item{NC}{a ncdf4 class object}

\item{sep}{character, used to separate dimensions if any, deafult is a space " "}
}
\value{
a character matrix of
 \itemize{
 \item{longname the name of the variable}
 \item{name the shortname of the variable}
 \item{unit the units - possibly ""}
 \item{dim the dimensions separated by underscores possibly ""}
}
}
\description{
Retrieve a matrix generic information about each variable
}
