% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{ncvarname_pretty}
\alias{ncvarname_pretty}
\title{Make a pretty list of variables "varname [dim1,dim2,...]"}
\usage{
ncvarname_pretty(nc, rm.names = TRUE)
}
\arguments{
\item{nc}{ncdf4 object}

\item{rm.names}{logical, if TRUE remove element names}
}
\value{
character vector of prettified varnames
}
\description{
Adapted from ncdf4::print.ncdf4()
}

