% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_latest_forecast}
\alias{query_latest_forecast}
\title{Query the latest forecast datasets}
\usage{
query_latest_forecast(ftime = c(NA, "latest", "0000", "0600", "1200",
  "1800")[2], ahead = c(NA, "000", "006", "084")[1])
}
\arguments{
\item{ftime}{character, the forecast time to retrieve
\itemize{
 \item{NA this disable filtering and gets the lot for the day}
 \item{'latest' get just the latest (the default)}
 \item{'0000', '0600', '1200''1800' one or more specific hours}
 }}

\item{ahead}{character, the 'forecast ahead' times as hhh.  This makes the
most sense when the value of ftime points to one time, such as 'latest', 
    '0600', etc.
\itemize{
\item{NA if NA, the default then all are returned}
\item{ any combination of 000, 001, 002, etc}
}}

\item{a}{list of DatasetRefClass, possibly empty or NULL}
}
\description{
Query the latest forecast datasets
}

