% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam_anl.R
\name{query_namanl}
\alias{query_namanl}
\title{Query NAM-ANL catalog for opendap resources by date}
\usage{
query_namanl(
  dates = c("2020-05-15", "2020-05-18"),
  ftime = c("0000", "0600", "1200", "1800"),
  ahead = c("000", "006"),
  threshold = namanl_threshold_date()
)
}
\arguments{
\item{dates}{Date or castable to Date, the date(s) to query}

\item{ftime}{character or integer, one or more 4 digit forecast hour (or castable to such).}

\item{ahead}{character or integer, one or more 4 digit forecast ahead hours (or castable to such)}

\item{threshold}{Date or castable to Date, the division date betwene "old" and "recent"}
}
\value{
one or more URL for .grb2 (opendap) resources, possibly including NA
}
\description{
Query NAM-ANL catalog for opendap resources by date
}
