% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam_anl.R
\name{query_namanl_catalog}
\alias{query_namanl_catalog}
\title{Query the catalog at the specified URL for the specified date}
\usage{
query_namanl_catalog(date, uri, pattern = "^.*0000_000\\\\.grb2$")
}
\arguments{
\item{date}{Date or character, cast-able to Date class}

\item{uri}{character, the base uri for the top level catalog}

\item{pattern}{character, one or more regular expression(s) used to match the opendap resources.}
}
\value{
relative URL for the resource or NA if not found
}
\description{
Query the catalog at the specified URL for the specified date
}
