% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam_cast.R
\name{query_namcast}
\alias{query_namcast}
\title{Query NAM-ANL catalog for opendap resources by date}
\usage{
query_namcast(
  dates = c("2020-05-15", "2020-05-18"),
  ftime = c("0000", "0600", "1200", "1800"),
  ahead = c("000", "006"),
  threshold = namcast_threshold_date()
)
}
\arguments{
\item{dates}{Date or castable to Date, the date(s) to query}

\item{ftime}{character or integer, 4 digit forecast statement hour(s) (or castable to such)
or "*" do include c("0000", "0600", "1200", "1800")}

\item{ahead}{character or integer, 4 digit forecast ahead hours (or castable to such)
or "*" to include all 000-084}

\item{threshold}{Date or castable to Date, the division date betwene "old" and "recent"}
}
\value{
one per input date, a URL for .grb2 (opendap) resources, possibly NA
}
\description{
Query NAM-ANL catalog for opendap resources by date
}
