% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_namcast_archive}
\alias{query_namcast_archive}
\title{Query the namcast resources for archived forecasts (different than analyses)}
\usage{
query_namcast_archive(day = "20150115", ftime = c(0, 6, 12, 18),
  ahead = c(0, 6), uri = top_url("archived-forecast"))
}
\arguments{
\item{day}{POSIXct or character in the form of 'YYYYmmdd', Default to today()}

\item{ftime}{character or numeric, on or more of the forecast time one or
more of [0, 6, 12, 18]}

\item{ahead}{3 character cycle timestamp ('000' now cast, '003 three hours ahead, etc)}

\item{uri, }{the base uri for NOMADS NCEP DODS NAM server}
}
\value{
character vector of uri
}
\description{
Query the namcast resources for archived forecasts (different than analyses)
}
