% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam_cast.R
\name{query_namcast_catalog}
\alias{query_namcast_catalog}
\title{Query the catalog at the specified URL for the specified date}
\usage{
query_namcast_catalog(
  date,
  uri,
  pattern = c("^.*0000_000\\\\.grb2$", "^.*0000_006\\\\.grb2$",
    "^.*0600_000\\\\.grb2$", "^.*0600_006\\\\.grb2$", "^.*1200_000\\\\.grb2$",
    "^.*1200_006\\\\.grb2$", "^.*1800_000\\\\.grb2$", "^.*1800_006\\\\.grb2$")
)
}
\arguments{
\item{date}{Date or character, cast-able to Date class}

\item{uri}{character, the base uri for the top level catalog}

\item{pattern}{character, one or more regular expressions used to match the
to just one of the opendap resources.}
}
\value{
relative URL for the resource or NA if not found
}
\description{
Query the catalog at the specified URL for the specified date
}
