% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib1.R
\name{read_grib1_inv}
\alias{read_grib1_inv}
\title{Read the grib1 inventory from a grib1 file}
\usage{

  read_grib1_inv(filename = "/mnt/ecocast/projects/devel/grib/namanl_218_20040303_0000_000.grb")
}
\arguments{
\item{filename}{the name of the grib1 file
@return a tibble of 0 or more rows}
}
\description{
Read the grib1 inventory from a grib1 file
}
