% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib1.R
\name{select_band}
\alias{select_band}
\title{Given a GRIB1 inventory table, select the band number}
\usage{
select_band(x, name = "TMP", level = "sfc")
}
\arguments{
\item{x}{data.frame of a GRIB1 inventory}

\item{name}{the variavle name}

\item{level}{the z level desired}
}
\value{
zero or more band numbers that match the name and level
}
\description{
Given a GRIB1 inventory table, select the band number
}
