% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{top_url}
\alias{top_url}
\title{Retrieve the top level catalog URL}
\usage{
top_url(
  what = c("analysis", "forecast", "archived-forecast", "analysis-historical")[1]
)
}
\arguments{
\item{what}{character specifies 'analysis' or 'forecast', "analysis-historical", 'current-forecast', 'archived-forecast'
'forecast' and 'archived-forecast' are the same thredds catalog but here for backward compatibility
'current-forecast' is actually a pointer NOMADS (not NCEP) which fuddles me
'analysis' points to a thredds catalog}
}
\value{
character URL
}
\description{
Retrieve the top level catalog URL
}
