% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib1.R
\name{which_grib}
\alias{which_grib}
\title{Determine by date whether to use grib1 or grib2}
\usage{
which_grib(
  date = c("2017-04-05", "2017-04-06"),
  origin = as.Date("2017-04-06")
)
}
\arguments{
\item{date}{one or more dates as character ("YYYY-mm-dd"), Date or POSIXt}

\item{origin}{the date at which NAM218 switches from grib1 to grib2}
}
\value{
character vector of length date with elements either 'grib1' or 'grib2'
}
\description{
Determine by date whether to use grib1 or grib2
}
