% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAM218.R
\name{which_var}
\alias{which_var}
\title{Retrieve the full variable name that best matches the provided pattern.}
\usage{
which_var(
  nc,
  pattern = c("Total_precipitation", "Convective_precipitation")[1],
  ignore.case = TRUE,
  ...
)
}
\arguments{
\item{nc}{a ncdf4 class object}

\item{pattern}{to match - see \code{grepl()}}

\item{ignore.case}{logical it TRUE ignore case in the pattern}

\item{...}{further arguments for \code{grepl()}}
}
\value{
character vector of zero or more matches
}
\description{
This functions helps resolve the issues that arise with changing GRIB2 variable
  names but note that it is possible for more than one variable name may be returned.
}
