% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAM218.R
\name{NAM218}
\alias{NAM218}
\title{Create an instance of NAM218RefClass}
\usage{
NAM218(nc, bb = NULL,
  proj = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{nc}{either a path, url, ncdf4 object or null}

\item{bb}{a 4-element vector of longlat as [xmin,xmax,ymin,ymax] defining the region of interest
If NULL then the extent of the dataset is used.}

\item{proj}{projection string for automatic transformation from NAM-ANL's native 'lcc' projection}
}
\value{
NAM218RefClass object
}
\description{
Create an instance of NAM218RefClass
}

