% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAM218.R
\name{NAM218_x_y_time}
\alias{NAM218_x_y_time}
\title{Retrieve a variable based upon [x,y,time] as a matrix}
\usage{
NAM218_x_y_time(X, name, bb = X$BB, xy = list(start = c(1, 1), count = c(-1,
  -1)), time_index = 1)
}
\arguments{
\item{X}{NAM218RefClass object}

\item{name}{character variable name}

\item{xy}{a two element list providing start and count values}

\item{time_index}{the default is to get just the first time index if there is}

\item{...}{further arguments (unused)}
}
\value{
numeric matrix
}
\description{
Retrieve a variable based upon [x,y,time] as a matrix
}

