% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namanl.R
\name{namanl_base_url}
\alias{namanl_base_url}
\title{Retrieve the base url for either HTML or OPENDAP}
\usage{
namanl_base_url(type = c("html", "thredds")[2])
}
\arguments{
\item{type}{character, either "html" or "thredds" (default)}
}
\value{
URL
}
\description{
Retrieve the base url for either HTML or OPENDAP
}
