% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{namanl_url}
\alias{namanl_url}
\title{Retrieve a URL for various resources}
\usage{
namanl_url(X, what = c("OPeNDAP", "OPeNDAP_form", "HTTPServer", "WCS", "WMS",
  "NetcdfServer")[1])
}
\arguments{
\item{X}{DatasetsRefClass object}

\item{what}{character - one of following
\itemize{
 \item{OPeNDAP for ncdf::nc_open()}
 \item{OPeNDAP_form for httr::BROWSE() OPeNDAP Dataset Access Form}
 \item{HTTPServer for utils::download.file()}
 \item{WCS not sure what one does with this}
 \item{WMS ditto}
 \item{NetcdfServer for httr::BROWSE() NetCDF Subset Service for Grids (interactive)}
 }}
}
\value{
character, URL
}
\description{
Retrieve a URL for various resources
}

