% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMcast.R
\name{NAMcast_x_y_lev_time}
\alias{NAMcast_x_y_lev_time}
\title{Retrieve a variable based upon [x,y,lev,time] as a matrix}
\usage{
NAMcast_x_y_lev_time(X, name, lev = 1000, time_index = 1, xy = list(start
  = c(1, 1), count = c(-1, -1)))
}
\arguments{
\item{X}{NAMcastRefClass object}

\item{name}{character variable name}

\item{lev}{numeric the level to extract}

\item{time_index}{the default is to get just the first time index if there is}

\item{xy}{a two element list providing start and count values}

\item{...}{further arguments (unused)}
}
\value{
numeric matrix
}
\description{
Retrieve a variable based upon [x,y,lev,time] as a matrix
}
