% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMcast.R
\name{NAMcastRefClass_get_time}
\alias{NAMcastRefClass_get_time}
\title{Compute the POSIXct timestamps}
\arguments{
\item{shift}{numeric, the raw day numbers and epoch conspire to cause a two day
difference from what is expected. So this param is used to correct that.}
}
\value{
POSIXct vector or NULL
}
\description{
As of summer 2020 time has been stored relative to the day of the forecast
embedded in the units of the time variable.  Prior to that it was relative to
0001-01-01 which required a 2-day shift to account for some mistakes made
prior to current conventions.
}
\seealso{
\url{https://github.com/dankelley/oce/issues/738}

\url{http://www.epic.noaa.gov/java/ncBrowse/mail/msg00059.html}
}
