% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_nam218}
\alias{query_nam218}
\title{Run a query for NAM-218 datasets}
\usage{
query_nam218(
  what = c("analysis", "forecast")[1],
  day = c("20060601", format(Sys.Date() - ifelse(what == "analysis", 4, 0), format =
    "\%Y\%m\%d"))[1],
  ftime = c(NA, "0000", "0600", "1200", "1800")[1],
  ahead = c(NA, "000", "003", "006", "084")[1]
)
}
\arguments{
\item{what}{character specifies 'analysis' or 'forecast'}

\item{day}{either POSIXct or an 8 character day (YYYYmmdd)}

\item{ftime}{4 character forecast period time stamp ('0000', '0006', etc.) or NA}

\item{ahead}{3 character cycle timestamp ('000' now cast, '003 three hours ahead, etc) or NA}
}
\value{
DatasetsRefClass or an empty list or NULL
}
\description{
Run a query for NAM-218 datasets
}
