% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam_anl.R
\name{query_namanl}
\alias{query_namanl}
\title{Query NAM-ANL catalog for opendap resources by date}
\usage{
query_namanl(
  dates = c("2020-05-15", "2020-05-18"),
  ftime = "0000",
  ahead = "000",
  threshold = namanl_threshold_date()
)
}
\arguments{
\item{dates}{Date or castable to Date, the date(s) to query}

\item{ftime}{character or integer, 4 digit forecast hour (or castable to such)}

\item{ahead}{character or integer, 4 digit forecast ahead hours (or castable to such)}

\item{threshold}{Date or castable to Date, the division date betwene "old" and "recent"}
}
\value{
one per input date, a URL for .grb2 (opendap) resources, possibly NA
}
\description{
Query NAM-ANL catalog for opendap resources by date
}
