% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_namcast}
\alias{query_namcast}
\title{Query the namcast resources}
\usage{
query_namcast(day = format(Sys.time(), "\%Y\%m\%d"), ftime = c(0, 6, 12,
  18), uri = "http://nomads.ncep.noaa.gov/dods/nam.xml")
}
\arguments{
\item{day}{POSIXct or character in the form of 'YYYYmmdd', Default to today()}

\item{ftime}{character or numeric, on or more of the forecast time one or
more of [0, 6, 12, 18]}

\item{charcater}{uri, the base uri for NOMADS NCEP DODS NAM server}
}
\value{
character vector of uri
}
\description{
Query the namcast resources
}

