% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_namcast}
\alias{query_namcast}
\title{Query the namcast resources for the current forecasts}
\usage{
query_namcast(
  day = format(Sys.Date(), "\%Y\%m\%d"),
  ftime = c(0, 6, 12, 18),
  pattern = "nam_[0-9]{2}z",
  uri = top_url("current-forecast")
)
}
\arguments{
\item{day}{POSIXct or character in the form of 'YYYYmmdd', Default to today()}

\item{ftime}{character or numeric, on or more of the forecast time one or
more of [0, 6, 12, 18]}

\item{pattern}{character of NULL, defaults to 'nam_[0-9]{2}z'}

\item{uri, }{the base uri for NOMADS NCEP DODS NAM server}
}
\value{
character vector of uri
}
\description{
Query the namcast resources for the current forecasts
}
